refactor: riskAssessmentDisplay 파일이동동
